Change appendix example responses to have access_token/Bearer and jwt/N_A per Mike's request and also fix up the associated text.
I disagree with the interruption that "Bearer" and "urn:ietf:params:oauth:token_type:jwt" together are inconsistent - a client can certainly know that the token is structured as JWT and still use it as an RFC 6750 bearer token.  But, while I don't think any change to the examples is necessary, it's not a hill I want to die on.